Use default port for MySQL if not specified in the connection secret
If a secret doesn't contain port, the provider won't be able to connect
as it will use an empty string, which then turns to 0. An attempt to
connect times out since this is not a valid port.

This is the case with connection secrets created by provider-aws, for
example for DBCluster. We should assume that the absence of the port
key/value means that connection should happen over the default port,
3306 for MySQL servers.